Rename artwork entity and update imports
