+++ src/models/dto/create-artwork.dto.ts
@@ -1,4 +1,4 @@
 import { PickType } from '@nestjs/swagger';
-import { Artwork } from '../entities/atrwork.entity';
+import { Artwork } from '../entities/artwork.entity';
 
 export class CreateArtworkDto extends PickType(Artwork, ['title', 'description']) {}
