+++ src/models/entities/exhibition.entity.ts
@@ -1,6 +1,6 @@
 import { Column, CreateDateColumn, Entity, JoinColumn, ManyToOne, OneToMany, PrimaryGeneratedColumn, UpdateDateColumn } from 'typeorm';
 import { User } from './user.entity';
-import { Artwork } from './atrwork.entity';
+import { Artwork } from './artwork.entity';
 import { ExhibitionArtwork } from './exhibition-artwork.entity';
 
 @Entity()
